config.py
import os
from dotenv import load_dotenv

# Chargement des variables d'environnement depuis un fichier .env si présent
load_dotenv()

# Configuration de Neo4j
NEO4J_URI = os.getenv("NEO4J_URI", "bolt://localhost:7687")
NEO4J_USER = os.getenv("NEO4J_USER", "neo4j")
NEO4J_PASSWORD = os.getenv("NEO4J_PASSWORD", "password")

# Configuration de l'application Flask
DEBUG = os.getenv("DEBUG", "True") == "True"
SECRET_KEY = os.getenv("SECRET_KEY", "dev_key_for_testing")