requirements.txt
flask==2.3.3
py2neo==2021.2.4
python-dotenv==1.0.0
flask-cors==4.0.0
pytest==7.4.0